+++ srcs/str/ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/24 18:41:43 by tmaluh            #+#    #+#             */
-/*   Updated: 2019/05/17 13:10:02 by tmaluh           ###   ########.fr       */
+/*   Updated: 2019/06/20 10:10:21 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-uchar	ft_strncmp(string s1, string s2, size_t n)
+int	ft_strncmp(string s1, string s2, size_t n)
 {
 	uchar	u1;
 	uchar	u2;
